requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==3.0.4
cmdstanpy==0.9.5
colorama==0.4.4
convertdate==2.3.0
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
elasticsearch==7.11.0
ephem==3.7.7.1
gluonts==0.6.5
graphviz==0.8.4
hijri-converter==2.1.1
holidays==0.10.5.2
idna==2.6
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
matplotlib==3.3.4
mxnet==1.6.0
numpy==1.20.1
pandas==1.2.1
parso==0.8.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.1.0
pmdarima==1.8.0
prompt-toolkit==3.0.14
pydantic==1.6.1
Pygments==2.7.4
PyMeeus==0.3.7
pyparsing==2.4.7
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pyzmq==22.0.2
requests==2.18.4
scikit-learn==0.24.1
scipy==1.6.0
setuptools-git==1.2
six==1.15.0
sklearn==0.0
statsmodels==0.12.2
threadpoolctl==2.1.0
toolz==0.11.1
torch==1.7.1
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.22
wcwidth==0.2.5
